third_party/blink/web_tests/fast/dom/length-attribute-mapping-expected.txt

This tests the mapping of length-type attributes to CSS length values.

<img>:
PASS: 90zz is mapped to 90px
PASS: 80% is mapped to 80%
PASS: 70%5 is mapped to 70%
PASS: 60%% is mapped to 60%
PASS: 50* is not mapped
PASS: 40*5 is not mapped
PASS: 30.5 is mapped to 30.5px

<col>:
PASS: 90zz is mapped to 90px
PASS: 80% is mapped to 80%
PASS: 70%5 is mapped to 70%
PASS: 60%% is mapped to 60%
PASS: 50* is not mapped
PASS: 40*5 is not mapped
PASS: 30.5 is mapped to 30.5px
